CMR-313: Mods-Mapping for VG-Wort-OpenKey Identifier https://jira.slub-dresden.de/browse/CMR-313
